requirements.txt
requests
python-whois
python-dotenv
pytest